// luna-frontend/src/api/type/homeData.types.ts
export interface HomeHeroData {
  title: string;
  highlight: string;
  text1: string;
  text2: string;
  image: string;
  imageAlt: string;
  buttons: {
    shop: string;
    story: string;
  };
}

export interface WhyLunaCard {
  img: string;
  title: string;
  text: string;
}

export interface AboutLunaData {
  title: string;
  highlight: string;
  description: string;
  buttonText: string;
  image: string;
  imageAlt: string;
}


export interface RitualData {
  title1: string;
  title2: string;
  description: string;
  buttonText: string;
}